Create the useCarousel hook
[Changes]
Create a useCarousel hook so that we can use it in multiple
parts of the code base. Currently, it is used in the amenities
carousel and the customer reviews section.
